refactor(catalog): prune public barrel, fix false dependencies and boundary violations
Remove common DTO re-exports that created phantom catalog dependency in
user-media and users modules. Prune 10 unused exports (23→13 symbols),
enforce consistent export type syntax, and fix test files that bypassed
the public API boundary.
