+++ apps/api/src/modules/ingestion/application/pipelines/snapshots.pipeline.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { SnapshotsPipeline } from './snapshots.pipeline';
 import { SnapshotsService, SnapshotBatchResult } from '../services/snapshots.service';
-import { MEDIA_REPOSITORY, SnapshotCandidate } from '../../../catalog/public';
+import { MEDIA_REPOSITORY, type SnapshotCandidate } from '../../../catalog/public';
 import { MediaType } from '../../../../common/enums/media-type.enum';
 
 describe('SnapshotsPipeline', () => {
